+++ shuup/core/defaults/order_statuses.py
@@ -5,15 +5,8 @@
 #
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
-from shuup.core.models import OrderStatus, OrderStatusRole
+from shuup.core.models import OrderStatusManager
 
 
 def create_default_order_statuses():
-    for i, props in enumerate([
-        {"name": u"received", "role": OrderStatusRole.INITIAL, "identifier": "recv", "default": True},
-        {"name": u"in progress", "identifier": "prog"},
-        {"name": u"complete", "role": OrderStatusRole.COMPLETE, "identifier": "comp", "default": True},
-        {"name": u"canceled", "role": OrderStatusRole.CANCELED, "identifier": "canc", "default": True}
-    ]):
-        if not OrderStatus.objects.filter(identifier=props["identifier"]).exists():
-            OrderStatus.objects.create(ordering=i, **props)
+    OrderStatusManager().ensure_default_statuses()